+++ packages/coin-support-btc/tests/__fixtures__/04.initializeTransaction.ts
@@ -34,6 +34,8 @@ export const valid: IInitializeTransactionTestCases[] = [
         outputs: [],
         isValidFee: true,
         isNotOverDustThreshold: false,
+        ownOutputAddressNotAllowed: [],
+        zeroAmountNotAllowed: false,
       },
     },
     mocks: {