+++ packages/coin-support-solana/src/operations/initializeTransaction/index.ts
@@ -19,6 +19,8 @@ export const initializeTransaction = async (
       outputs: [],
       hasEnoughBalance: true,
       isValidFee: true,
+      ownOutputAddressNotAllowed: [],
+      zeroAmountNotAllowed: false,
     },
     userInputs: {
       outputs: [],