+++ packages/coin-support-tron/src/utils/tronweb.ts
@@ -1,0 +1,12 @@
+let tronWebInstance: any | undefined;
+
+export const getCoinSupportTronWeb = () => {
+  if (!tronWebInstance) {
+    throw new Error('tronweb has not been set yet');
+  }
+  return tronWebInstance;
+};
+
+export const setCoinSupportTronWeb = (tronweb: any) => {
+  tronWebInstance = tronweb;
+};