+++ packages/coin-support/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @cypherock/coin-support
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [[`c02dfec5`](https://github.com/Cypherock/cypherock-cysync/commit/c02dfec50489c172b7086f8e2b7a3020cfd43581)]:
+  - @cypherock/coin-support-interfaces@0.0.4
+  - @cypherock/coin-support-solana@0.0.6
+  - @cypherock/coin-support-near@0.0.6
+  - @cypherock/coin-support-btc@0.0.6
+  - @cypherock/coin-support-evm@0.0.6
+  - @cypherock/cysync-utils@0.0.5
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [[`3c21df54`](https://github.com/Cypherock/cypherock-cysync/commit/3c21df54f01f856623c0531fc43a8b66eb37421b)]:
+  - @cypherock/coins@0.0.5
+  - @cypherock/coin-support-btc@0.0.5
+  - @cypherock/coin-support-evm@0.0.5
+  - @cypherock/coin-support-near@0.0.5
+  - @cypherock/coin-support-solana@0.0.5
+
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`d60fece4`](https://github.com/Cypherock/cypherock-cysync/commit/d60fece42cd3ae968614b80ca4c5886cdcf2614a), [`e1d692d8`](https://github.com/Cypherock/cypherock-cysync/commit/e1d692d8b847b620ef4149ae0f43b82bfcc54735), [`6555bece`](https://github.com/Cypherock/cypherock-cysync/commit/6555bece4751f8cfe02b3a35ee7541bec935d4d0), [`c25efafc`](https://github.com/Cypherock/cypherock-cysync/commit/c25efafcee92cc1c9f1eac51f0641edddff73b10)]:
+  - @cypherock/cysync-utils@0.0.4
+  - @cypherock/coins@0.0.4
+  - @cypherock/coin-support-solana@0.0.4
+  - @cypherock/coin-support-near@0.0.4
+  - @cypherock/coin-support-btc@0.0.4
+  - @cypherock/coin-support-evm@0.0.4
+
 ## 0.0.3
 
 ### Patch Changes